Rework the ClusterContentLibraryItem controller tests
Parts of the prior tests made the conversion to v1a2 take longer than
it really should have due to some of the subtleness what, where, & how
the expected values are coming from, and what is mismatched.

Leave the ContentLibraryItem controller as-is for now to let this
commit stew for a bit.