Range over the CreateOrUpdate channel
The async could return multiple errors - vim error and
k8s error - so emit an event for all.

This is kind of ugly in order to still emit an event in
the non-error case when the channel is just closed. Alt
would be to like send 'nil' on the channel for success
I guess.