Do not requeue VMIC for empty Spec fields
If these empty Spec fields are updated, the Watch() will cause the
controller to reconcile the object.

Add NoRequeueError that returns the controller-runtime TerminalError
that does not do a requeue but is logged and counted as an error.

While here, update the VMIC ReconcileNormal() to use the logger
from the context instead of parent reconciler so the object name
is still there.